# apps/smart_home_microservices/telemetry_service/telemetry_service_openapi.json
{"openapi":"3.1.0","info":{"title":"Telemetry Service","description":"Сервис телеметрии: абстрактный роутер","version":"0.1.0"},"paths":{"/health_check":{"get":{"summary":"Health Check","description":"Проверка состояния сервиса","operationId":"health_check_health_check_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Response Health Check Health Check Get"}}}}}}},"/measurement_processes":{"get":{"summary":"Get Measurement Processes","description":"Вернуть все процессы измерений, которые обрабатывал этот сервис\n(с возможностью фильтрации по датчикам и времени запуска).","operationId":"get_measurement_processes","parameters":[{"name":"active_only","in":"query","required":false,"schema":{"type":"boolean","default":false,"title":"Active Only"}},{"name":"min_start_ts","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Min Start Ts"}},{"name":"max_start_ts","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Max Start Ts"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Body_get_measurement_processes"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MeasurementProcessInfo"},"title":"Response Get Measurement Processes"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/measurement_process/{meas_proc_id}":{"get":{"summary":"Get Measurement Process","description":"Вернуть информацию об отдельном процессе измерений по его ID","operationId":"get_measurement_process_measurement_process__meas_proc_id__get","parameters":[{"name":"meas_proc_id","in":"path","required":true,"schema":{"type":"integer","title":"Meas Proc Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MeasurementProcessInfo"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/measurement_process/{meas_proc_id}/samples":{"get":{"summary":"Get Samples","description":"Достать из БД все измерения для конкретного процесса.\nОпционально -- можно извлечь не все измерения, а только в заданном диапазоне времени.","operationId":"get_samples_measurement_process__meas_proc_id__samples_get","parameters":[{"name":"meas_proc_id","in":"path","required":true,"schema":{"type":"integer","title":"Meas Proc Id"}},{"name":"start_ts","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Start Ts"}},{"name":"end_ts","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"End Ts"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TelemetrySampleInfo"},"title":"Response Get Samples Measurement Process  Meas Proc Id  Samples Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Body_get_measurement_processes":{"properties":{"sensor_ids":{"anyOf":[{"type":"integer"},{"items":{"type":"integer"},"type":"array"},{"type":"null"}],"title":"Sensor Ids"},"sensor_uuids":{"anyOf":[{"type":"string","format":"uuid"},{"items":{"type":"string","format":"uuid"},"type":"array"},{"type":"null"}],"title":"Sensor Uuids"}},"type":"object","title":"Body_get_measurement_processes"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"MeasurementProcessInfo":{"properties":{"sensor_id":{"type":"integer","title":"Sensor Id","description":"Идентификатор датчика, к которому относится процесс измерения"},"sample_format":{"$ref":"#/components/schemas/TelemetrySampleFormat","description":"Формат данных, в котором датчик возвращает измерения"},"started_at":{"type":"string","format":"date-time","title":"Started At","description":"Время запуска процесса измерения"},"stopped_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Stopped At","description":"Время остановки процесса измерения (None, если процесс ещё выполняется)"},"sampling_interval":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Sampling Interval","description":"Интервал между измерениями в секундах. Может быть None, если датчик сам посылает данные по необходимости"}},"type":"object","required":["sensor_id","sample_format"],"title":"MeasurementProcessInfo","description":"Информация о процессе измерения, запущенном на датчике.\nНа одном датчике может быть запущено несколько процессов измерения одновременно."},"TelemetrySampleFormat":{"type":"string","enum":["CUSTOM","FLOAT_WITH_TIMESTAMP","FLOAT_BINARY","INT","BOOL","TEXT"],"title":"TelemetrySampleFormat","description":"В каком формате заливать телеметрию в очередь.\nNB: FLOAT_WITH_TIMESTAMP и FLOAT потом будует конвертироваться в один и тот же класс FloatTelemetrySample)"},"TelemetrySampleInfo":{"properties":{"timestamp":{"type":"string","format":"date-time","title":"Timestamp"},"value":{"anyOf":[{},{"type":"null"}],"title":"Value"}},"type":"object","title":"TelemetrySampleInfo","description":"Базовый класс для отдельных измерений телеметрии в процессе измерений."},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}